changes
re-worked illuminance suppression logic:
now nothing is reported if presence change was suppressed by luminance
level. Suppression is checked on detection only. So if it's suppressed -
it stays that way until the next off->on transition if illuminance
conditions are ok

initial start: performing reads from pins for mmWave and PIR presence to
properly initialize the first state